Add populate_metrics_from_database function Remove save to file and watcher Remove files / rename mrg-dump-load to mrg-load.c Update CMakeLists.txt
