+++ src/daemon/daemon-shutdown-watcher.h
@@ -25,7 +25,6 @@ typedef enum {
     WATCHER_STEP_ID_STOP_DBENGINE_TIERS,
     WATCHER_STEP_ID_STOP_METASYNC_THREADS,
     WATCHER_STEP_ID_CLOSE_SQL_DATABASES,
-    WATCHER_STEP_ID_MRG_SAVE,
     WATCHER_STEP_ID_REMOVE_PID_FILE,
     WATCHER_STEP_ID_FREE_OPENSSL_STRUCTURES,
